refactor: CollectionDetail 관련 로직 리팩터링 및 매퍼 분리
- `CollectionDetailMapper`를 신설하여 DTO에서 Model로의 변환 로직을 분리
- `CollectionDetailModelNew`에서 `userId` 및 `isMine` 관련 로직 제거
- `CollectionRepository`에서 `PreferencesManager` 의존성을 제거하고 매퍼를 사용하도록 수정
- `CollectionDetailScreen` 및 관련 컴포넌트에서 `isMine` 파라미터 제거 및 노출 로직 수정
