+++ app/src/main/java/com/flint/domain/model/collection/CollectionDetailModel.kt
@@ -1,11 +1,8 @@
 package com.flint.domain.model.collection
 
-import com.flint.data.dto.collection.response.CollectionDetailResponseDto
 import com.flint.domain.model.AuthorModelNew
 import com.flint.domain.model.content.ContentModelNew
-import com.flint.domain.type.UserRoleType
 import kotlinx.collections.immutable.ImmutableList
-import kotlinx.collections.immutable.toImmutableList
 import java.time.LocalDate
 
 data class CollectionDetailModelNew(
@@ -17,45 +14,4 @@ data class CollectionDetailModelNew(
     val thumbnailUrl: String,
     val isBookmarked: Boolean,
     val title: String,
-    private val userId: String,
-) {
-    constructor(
-        collectionDetail: CollectionDetailResponseDto,
-        userId: String,
-    ) : this(
-        author = collectionDetail.author.toModel(),
-        contents = collectionDetail.contents.map { it.toModel() }.toImmutableList(),
-        createdAt = LocalDate.parse(collectionDetail.createdAt),
-        description = collectionDetail.description,
-        id = collectionDetail.id,
-        thumbnailUrl = collectionDetail.thumbnailUrl,
-        isBookmarked = collectionDetail.isBookmarked,
-        title = collectionDetail.title,
-        userId = userId
-    )
-
-    val isMine: Boolean = author.id == userId
-}
-
-private fun CollectionDetailResponseDto.Author.toModel(): AuthorModelNew {
-    return AuthorModelNew(
-        id = id,
-        nickname = nickname,
-        profileImageUrl = profileImageUrl,
-        userRole = runCatching { UserRoleType.valueOf(userRole) }.getOrDefault(UserRoleType.NONE)
-    )
-}
-
-private fun CollectionDetailResponseDto.Content.toModel(): ContentModelNew {
-    return ContentModelNew(
-        director = director,
-        bookmarkCount = bookmarkCount,
-        id = id,
-        isBookmarked = isBookmarked,
-        isSpoiler = isSpoiler,
-        reason = reason,
-        imageUrl = imageUrl,
-        title = title,
-        year = year,
-    )
-}
+)
