+++ app/src/main/java/com/flint/domain/repository/CollectionRepository.kt
@@ -1,22 +1,18 @@
 package com.flint.domain.repository
 
-import com.flint.core.common.util.DataStoreKey.USER_ID
 import com.flint.core.common.util.suspendRunCatching
 import com.flint.data.api.CollectionApi
 import com.flint.data.dto.collection.request.CollectionCreateRequestDto
 import com.flint.data.dto.collection.response.CollectionDetailResponseDto
-import com.flint.data.local.PreferencesManager
 import com.flint.domain.mapper.collection.toModel
 import com.flint.domain.model.collection.CollectionCreateModel
 import com.flint.domain.model.collection.CollectionDetailModelNew
 import com.flint.domain.model.collection.CollectionListModel
 import com.flint.domain.model.collection.CollectionsModel
-import kotlinx.coroutines.flow.first
 import javax.inject.Inject
 
 class CollectionRepository @Inject constructor(
     private val apiService: CollectionApi,
-    private val preferencesManager: PreferencesManager,
 ) {
     // 컬렉션 목록 조회 (페이지네이션)
     suspend fun getCollections(cursor: Long?, size: Int): Result<CollectionsModel> =
@@ -29,7 +25,7 @@ class CollectionRepository @Inject constructor(
 
     // 컬렉션 생성
     suspend fun postCollectionCreate(
-        requestDto: CollectionCreateRequestDto
+        requestDto: CollectionCreateRequestDto,
     ): Result<CollectionCreateModel> =
         suspendRunCatching {
             apiService.postCollectionCreate(requestDto).data.toModel()
@@ -41,9 +37,8 @@ class CollectionRepository @Inject constructor(
         suspendRunCatching {
             val response: CollectionDetailResponseDto =
                 apiService.getCollectionDetail(collectionId).data
-            val userId: String = preferencesManager.getString(USER_ID).first()
 
-            CollectionDetailModelNew(response, userId)
+            response.toModel()
         }
 
     // 최근 본 컬렉션 목록 조회
